+++ specutils/io/default_loaders/apogee.py
@@ -24,51 +24,62 @@ __all__ = ['apVisit_identify', 'apStar_identify', 'aspcapStar_identify',
 
 def apVisit_identify(origin, *args, **kwargs):
     """
-    Check whether given filename is FITS. This is used for Astropy I/O
+    Check whether given file is FITS. This is used for Astropy I/O
     Registry.
     """
-    return (isinstance(args[0], str) and
-            args[0].startswith('apVisit') and
-            fits.connect.is_fits(origin, *args))
+    with fits.open(args[0]) as hdulist:
+        # Test if fits has extension of type BinTable and check for
+        # apVisit-specific keys
+        return (hdulist[0].header.get('SURVEY') == 'apogee' and
+                len(hdulist) > 4 and
+                hdulist[1].header.get('BUNIT', 'none').startswith('Flux') and
+                hdulist[2].header.get('BUNIT', 'none').startswith('Flux') and
+                hdulist[4].header.get('BUNIT', 'none').startswith('Wavelen'))
 
 
 def apStar_identify(origin, *args, **kwargs):
     """
-    Check whether given filename is FITS. This is used for Astropy I/O
+    Check whether given file is FITS. This is used for Astropy I/O
     Registry.
     """
-    return (isinstance(args[0], str) and
-            args[0].lower().split('.')[-1] == 'fits' and
-            args[0].startswith('apStar'))
+    with fits.open(args[0]) as hdulist:
+        # Test if fits has extension of type BinTable and check for
+        # apogee-specific keys + keys for individual apVisits
+        return (hdulist[0].header.get('SURVEY') == 'apogee' and
+                hdulist[0].header.get('SFILE1', 'none').startswith('apVisit'))
 
 
 def aspcapStar_identify(origin, *args, **kwargs):
     """
-    Check whether given filename is FITS. This is used for Astropy I/O
+    Check whether given file is FITS. This is used for Astropy I/O
     Registry.
     """
-    return (isinstance(args[0], str) and
-            args[0].lower().split('.')[-1] == 'fits' and
-            args[0].startswith('aspcapStar'))
+    with fits.open(args[0]) as hdulist:
+        # Test if fits has extension of type BinTable and check for
+        # aspcapStar-specific keys
+        return (hdulist[0].header.get('TARG1') is not None and
+                len(hdulist) > 4 and
+                hdulist[1].header.get('NAXIS1', 0) > 8000 and
+                hdulist[2].header.get('NAXIS1', 0) > 8000 and
+                hdulist[-1].header.get('TTYPE45') == 'ASPCAPFLAG')
 
 
 @data_loader(label="APOGEE apVisit", identifier=apVisit_identify, extensions=['fits'])
-def apVisit_loader(file_name, **kwargs):
+def apVisit_loader(file_obj, **kwargs):
     """
     Loader for APOGEE apVisit files.
 
     Parameters
     ----------
-    file_name: str
-        The path to the FITS file
+    file_obj: str or file-like
+        FITS file name or object (provided from name by Astropy I/O Registry).
 
     Returns
     -------
     data: Spectrum1D
         The spectrum that is represented by the data in this table.
     """
-    name = os.path.basename(file_name.rstrip(os.sep)).rsplit('.', 1)[0]
-    hdulist = fits.open(file_name, **kwargs)
+    hdulist = fits.open(file_obj, **kwargs)
 
     header = hdulist[0].header
     meta = {'header': header}
@@ -99,22 +110,21 @@ def apVisit_loader(file_name, **kwargs):
 
 
 @data_loader(label="APOGEE apStar", identifier=apStar_identify, extensions=['fits'])
-def apStar_loader(file_name, **kwargs):
+def apStar_loader(file_obj, **kwargs):
     """
     Loader for APOGEE apStar files.
 
     Parameters
     ----------
-    file_name: str
-        The path to the FITS file
+    file_obj: str or file-like
+        FITS file name or object (provided from name by Astropy I/O Registry).
 
     Returns
     -------
     data: Spectrum1D
         The spectrum that is represented by the data in this table.
     """
-    name = os.path.basename(file_name.rstrip(os.sep)).rsplit('.', 1)[0]
-    hdulist = fits.open(file_name, **kwargs)
+    hdulist = fits.open(file_obj, **kwargs)
 
     header = hdulist[0].header
     meta = {'header': header}
@@ -141,22 +151,21 @@ def apStar_loader(file_name, **kwargs):
 
 
 @data_loader(label="APOGEE aspcapStar", identifier=aspcapStar_identify, extensions=['fits'])
-def aspcapStar_loader(file_name, **kwargs):
+def aspcapStar_loader(file_obj, **kwargs):
     """
     Loader for APOGEE aspcapStar files.
 
     Parameters
     ----------
-    file_name: str
-        The path to the FITS file
+    file_obj: str or file-like
+        FITS file name or object (provided from name by Astropy I/O Registry).
 
     Returns
     -------
     data: Spectrum1D
         The spectrum that is represented by the data in this table.
     """
-    name = os.path.basename(file_name.rstrip(os.sep)).rsplit('.', 1)[0]
-    hdulist = fits.open(file_name, **kwargs)
+    hdulist = fits.open(file_obj, **kwargs)
 
     header = hdulist[0].header
     meta = {'header': header}