+++ specutils/io/default_loaders/wcs_fits.py
@@ -18,22 +18,18 @@ def identify_wcs1d_fits(origin, *args, **kwargs):
     # check if file can be opened with this reader
     # args[0] = filename
     with fits.open(args[0]) as hdu:
-        if (
-            # check if number of axes is one
-            hdu[0].header['NAXIS'] == 1 and
-            hdu[0].header.get('WCSDIM', 1) == 1 and
-            'WAT1_001' not in hdu[0].header and
-            # check if CTYPE1 kep is in the header
-            'CTYPE1' in hdu[0].header
-            ):
-            return True
+        # check if number of axes is one
+        return (hdu[0].header['NAXIS'] == 1 and
+                hdu[0].header.get('WCSDIM', 1) == 1 and
+                # check in CTYPE1 key for linear solution
+                hdu[0].header.get('CTYPE1', '').upper() != 'MULTISPEC')
 
     return False
 
 
 @data_loader("wcs1d-fits", identifier=identify_wcs1d_fits,
              dtype=Spectrum1D, extensions=['fits'])
-def wcs1d_fits_loader(file_name, spectral_axis_unit=None, flux_unit=None,
+def wcs1d_fits_loader(file_obj, spectral_axis_unit=None, flux_unit=None,
                       hdu_idx=0, **kwargs):
     """
     Loader for single spectrum-per-HDU spectra in FITS files, with the spectral
@@ -43,8 +39,8 @@ def wcs1d_fits_loader(file_name, spectral_axis_unit=None, flux_unit=None,
 
     Parameters
     ----------
-    file_name : str
-        The path to the FITS file.
+    file_obj: str or file-like
+        FITS file name or object (provided from name by Astropy I/O Registry).
     spectral_axis_unit: str or `~astropy.Unit`, optional
         Units of the spectral axis. If not given (or None), the unit will be
         inferred from the CUNIT in the WCS.  Not that if this is providded it
@@ -62,7 +58,7 @@ def wcs1d_fits_loader(file_name, spectral_axis_unit=None, flux_unit=None,
     """
     logging.info("Spectrum file looks like wcs1d-fits")
 
-    with fits.open(file_name, **kwargs) as hdulist:
+    with fits.open(file_obj, **kwargs) as hdulist:
         header = hdulist[hdu_idx].header
         wcs = WCS(header)
 
@@ -90,16 +86,17 @@ def identify_iraf_wcs(origin, *args):
     The difference of this with respect to wcs1d is that this can work with
     WCSDIM == 2
     """
-    return (fits.connect.is_fits(origin, *args) and
-            'WAT1_001' in fits.getheader(args[0]) and
-            not (fits.getheader(args[0])['TELESCOP'] == 'SDSS 2.5-M' and
-                 fits.getheader(args[0])['FIBERID'] > 0)
-            )
+    with fits.open(args[0]) as hdulist:
+        return ('WAT1_001' in hdulist[0].header and not
+                (hdulist[0].header['TELESCOP'] == 'SDSS 2.5-M' and
+                 hdulist[0].header['FIBERID'] > 0) and
+                (hdu[0].header.get('WCSDIM', 1) > 1 or
+                 hdu[0].header.get('CTYPE1', '').upper() == 'MULTISPEC'))
 
 
 @data_loader('iraf', identifier=identify_iraf_wcs, dtype=Spectrum1D,
              extensions=['fits'])
-def non_linear_wcs1d_fits(file_name, spectral_axis_unit=None, flux_unit=None,
+def non_linear_wcs1d_fits(file_obj, spectral_axis_unit=None, flux_unit=None,
                           **kwargs):
     """Read wcs from files written by IRAF
 
@@ -109,8 +106,8 @@ def non_linear_wcs1d_fits(file_name, spectral_axis_unit=None, flux_unit=None,
     Parameters
     ----------
 
-    file_name : str
-        Name of file to load
+    file_obj: str or file-like
+        FITS file name or object (provided from name by Astropy I/O Registry).
 
     spectral_axis_unit : `~astropy.Unit`, optional
         Spectral axis unit, default is None in which case will search for it
@@ -128,7 +125,7 @@ def non_linear_wcs1d_fits(file_name, spectral_axis_unit=None, flux_unit=None,
 
     logging.info('Loading 1D non-linear fits solution')
 
-    with fits.open(file_name, **kwargs) as hdulist:
+    with fits.open(file_obj, **kwargs) as hdulist:
         header = hdulist[0].header
         for wcsdim in range(1, header['WCSDIM'] + 1):
             ctypen = header['CTYPE{:d}'.format(wcsdim)]